# Optimize schedules query